fix: GPU metrics u128→u64, ptx_map head_dim div-by-zero, rosetta dimension overflow (Refs #496)
- serve: cast as_millis() to u64 for consistent JSON type with other
  handler paths
- ptx_map: guard head_dim calculation against num_heads=0 on corrupt
  models
- rosetta: use checked_mul for dimension product to prevent silent
  overflow on large tensors

+++ crates/apr-cli/src/commands/serve/handler_gpu_completion.rs
@@ -323,7 +323,7 @@ async fn handle_gpu_chat_completion(
                 "completion_tokens": tokens_generated,
                 "total_tokens": input_tokens.len() + tokens_generated
             },
-            "_apr_metrics": {"latency_ms": start.elapsed().as_millis(), "tok_per_sec": tok_per_sec}
+            "_apr_metrics": {"latency_ms": start.elapsed().as_millis() as u64, "tok_per_sec": tok_per_sec}
         }))
         .into_response()
     }
